fixed a bug in llvm-readobj introduced by my earlier changes
details: the record counts I added to each function entry were 32 bit integers
because the total number of records is limited by that size integer.
the problem with that decision lies with the StackMapParser's way of computing
the padding inbetween each record: it rounds up the size of every record to
be divisible by 8. This only works as a way to offset to the next adjacent aligned record
if you assume the first stackmap record started at an 8-byte aligned boundary,
but it is not true anymore after adding the 32 bit integer count just before it.

Thus, this patch just widens the count size to restore the assumption made by the stackmap parser
to solve the problem